Change character from where to include response syntax
This is to allow more blocks of text in the response. The previous start
for body syntax is starting the region from the first empty line to the
end of file. This is fine for a response with only two blocks of text,
so I'm inserting a special UTF-8 space to start inner syntax from there.

Also refresh syntax after pasting a content.